UBUNTU: [arm] Fix kexec on ARM by properly calling the relocation function
The kexec syscall function is broken on ARM due to it not properly calling
the relocation stub with the correct arguments. This patch puts machine_kexec
in line with the other architectures, and allows kexec to work peroply on ARM.
It has been tested on the versatile kernel successfully.

Signed-off-by: Michael Casadevall <michael.casadevall@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>
(cherry picked from commit 6cd8dd766aec17ce66aa4f732928d05534da0937)