UBUNTU: SAUCE: (drop after 2.6.28) [PATCH] [eCryptfs] Regression in unencrypted filename symlinks
Bug: #322532

The addition of filename encryption caused a regression in symlinks when
filename encryption is not being used. readlink() returned n chars,
while strlen() on the returned symlink target reported n - 1 chars. Most
applications didn't notice since the extra char was '/0', but a version
control system (bzr) caught the bug.

This patch is on it's way upstream - rtg

Signed-off-by: Tyler Hicks <tyhicks@linux.vnet.ibm.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>