UBUNTU: SAUCE: (remove after 2.6.28) V4L/DVB (10216): saa7127: fix broken S-Video with saa7129
Bug: #329267

Register 0x2d has to be set differently in the saa7129 compared to the
saa7127. This was not done correctly, so S-Video was broken in certain
circumstances.

This fixes a regression introduced in 2.6.28.

This the patch as attached to http://article.gmane.org/gmane.linux.drivers.video-input-infrastructure/166,
which is a backport of the upstream commit.

Signed-off-by: Martin Dauskardt <martin.dauskardt@gmx.de>
Signed-off-by: Hans Verkuil <hverkuil@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>