UBUNTU: SAUCE: cpufreq-nforce2: probe failures are not errors
Bug: #332170

When we cpufreq-nforce2 fails to find the appropriate hardware it
emits an error to indicate this.  However when this driver is built
into the kernel this probe and fail on all non-nforce2 systems.  This
leads us to see the following error between the BIOS and usplash:

    [    1.446954] cpufreq: No nForce2 chipset.

This error is purely informational and should not be at KERN_ERR.  Move to
this to KERN_INFO.

A corresponding fix is already moving upstream and expected to hit 2.6.30.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>