UBUNTU: link-headers -- only link directories which do not already exist
Bug: #315252

When building the flavour specific header link-balls we ensure we have
all of the directories linked by iterating over those.  However when
the directory already exists due to a local file within it we do not
notice the directory and link it regardless.  This leaves us with
stuttered links on install:

  /usr/src/linux-headers-2.6.28-8-generic/include/linux/linux
  /usr/src/linux-headers-2.6.28-8-generic/include/asm-x86/asm-x86

This occurs because we check that the destination does not already
exist before linking, but we only check that it is not a file.  Existing
directories are not matched and are relinked.  Fix this to detect both.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>