ARM.imx51 Freescale:ENGR00108843 ipu3: support bigger hsync widths
The hsync pulse width was based on the DI clock and would easily overflow
the 9-bit field. This problem was seen with standard VESA timings. Changing
the waveform counter to be based on the pixel clock instead.

Also, add some checks for overflowing the register fields of the DI counters.

Signed-off-by: Rob Herring <r.herring@freescale.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>