UBUNTU: Fix the VFP handling on the Feroceon CPU
This CPU generates synchronous VFP exceptions in a non-standard way -the
FPEXC.EX bit set but without the FPSCR.IXE bit being set like in the VFP
subarchitecture 1 or just the FPEXC.DEX bit like in VFP subarchitecture
2. The main problem is that the faulty instruction will be re-executed
indefinitely without being emulated. This patch ensures that the
VFP exception is treated as synchronous.

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>