UBUNTU: SAUCE: (drop after 2.6.28) i2c-piix4: Add support to SB800 SMBus changes
Bug: #338108

Add support for the AMD SB800 Family series of products.
Major changes include the changes to addressing the SMBus registers at different
location from the locations in the previous compatible parts from AMD such as
SB400/SB600/SB700. For SB800, the main features and register definitions of
SMBus and other interfaces are still compatible with the previous products with
the only change being in how to access the internal registers for these blocks.

This patch should appear in the 2.6.30 merge window. Normally I would not take a
patch like this unless it is in an upstream maintainer repository, but jdelvare
does not appear to use git. Besides, the patch is quite self contained and has no
chance of impacting systems that do not contain the SB00 PCI device. - rtg

http://www.kernel.org/pub/linux/kernel/people/jdelvare/linux-2.6/jdelvare-i2c/i2c-piix4-add-support-for-sb800.patch

Signed-off-by: Shane Huang <shane.huang@amd.com>
Signed-off-by: Jean Delvare <khali@linux-fr.org>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>