UBUNTU: SAUCE: Fixing symbol name in HECI module
Bug: #336549

The HECI module define a switch of debug message, which symbol name
conflicts with another symbol in "arch/x86/kernel/entry_32.S".
Change the symbol name from "debug" to "heci_debug" so that the compile
wont fail when "CONFIG_HECI=y".

This bug appears in the HECI version 3.2.0.24 but has been fixed in the
next version (5.0.0.30). But there are a lot of modification between
these two version.

Recommend for Intrepid, Jaunty and Karmic, and drop off if we update
the module to next verion or later.

Signed-off-by: Ike Panhc <ike.pan@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>