UBUNTU: SAUCE: ACPI: Populate DIDL before registering ACPI video device on Intel
BugLink: http://bugs.launchpad.net/bugs/228399

commit 74a365b3f354fafc537efa5867deb7a9fadbfe27 backported

Note we lose the KMS support which makes most of this redundant
including the DIDL update.

Intel graphics hardware that implements the ACPI IGD OpRegion spec
requires that the list of display devices be populated before any ACPI
video methods are called. Detect when this is the case and defer
registration until the opregion code calls it. Fixes crashes on HP
laptops.

http://bugzilla.kernel.org/show_bug.cgi?id=11259

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Acked-by: Eric Anholt <eric@anholt.net>
Signed-off-by: Len Brown <len.brown@intel.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>