UBUNTU: Build-in "Ram block device support" to boot up with initramfs
Bug: #329098

Debian kernel for iop32x can boot up with initramfs cause it was configurated
to build-in "Ram block device support", while our Jaunty kernel build it as a
module. So Jaunty kernel for iop32x can not boot up with initramfs as the bug
tracker recorded.

This patch fixes this bug by changing the config to build-in "Ram block device
support" for Jaunty iop32x kernel config.

Signed-off-by: Bryan Wu <bryan.wu@canonical.com>
Signed-off-by: Amit Kucheria <amit.kucheria@canonical.com>