UBUNTU: SAUCE: acer: rfkill disable quirk for ACER Aspire One
Bug: #319825

We have recently added software polled rfkill support for the Acer laptops.
However for Acer Aspire One WMI interface is non-functional which prevents
the acer-wmi module from correctly determining the current rfkill state.
This leads the kernel to disable wireless permenantly as it believes the
kill switch is enabled.  Add support for a hardware rfkill quirk flag and
add a quirk for the Acer Aspire One models we know about.  Also adds an
rfkill=Y/N flag which can be used to override the quirk.  Also cleanup
the failure cleanup handling for brightness and email led.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>