UBUNTU: SAUCE: md: wait for possible pending deletes after stopping an array
Bug: #334994

When stopping an array, we first wait for any pending deletes to occur
and then remove all the disks from the array.  Unfortuanatly it is the
removal of the disks from the array which triggers the the pending deletes.
We should be waiting for them after they have been started.  Without this
it is possible to stop and then attempt to restart an array before the
deletes have occured leading to a start failure.  This is a common pattern
if an array is found to be degraded on boot.  This leads to boot failures
when failing back to degraded mode is acceptable.

Reorder the wait for completion till after the delete.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>