UBUNTU: [Jaunty] Change LPIA configuration to compile with CONFIG_NETFILTER_XT_MATCH_RECENT
Bug: #355291

The 'recent' module of iptables is broken on lpia because the kernel is
compiled without CONFIG_NETFILTER_XT_MATCH_RECENT. This is a regression
over Intrepid:

$ grep RECENT ./config-2.6.2*
./config-2.6.27-4-lpia:CONFIG_IP_NF_MATCH_RECENT=m
./config-2.6.28-11-lpia:# CONFIG_NETFILTER_XT_MATCH_RECENT is not set

$ cat /proc/version_signature
Ubuntu 2.6.28-11.40-lpia
$ sudo iptables -A INPUT -p tcp --dport 22 -m state --state NEW -m
recent --set
iptables: No chain/target/match by that name

Ufw uses this module when using the LIMIT command, which causes the
firewall to not load on boot due to iptables-restore failing. Ufw users
are only affected when using LIMIT rules.

The LPIA configuration was modifified to match the NETFILTER
configuration for mad64 and i386.

Signed-off-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>