UBUNTU: disable CONFIG_SND_HDA_INPUT_BEEP on amd64 and i386
Bug: #331589

It appears that different hda codecs have differing implementations of the
beep code, some offering volume controls to adjust/mute the beep, and others
offering no controls. Since many users find this beep annoying, and not all
codecs allow the beep volume to be controlled, lets get rid of it for now.

Signed-off-by: Luke Yelavich <luke.yelavich@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>