UBUNTU: [Config] Disable libusual and the ub driver
Bug: #364538
BugLink: https://bugs.launchpad.net/ubuntu/+source/linux/+bug/364538
BumpABI: yes

ABI changes:
- Missing symbols:
    MISS : usb_usual_check_type
    MISS : storage_usb_ids
    MISS : usb_usual_set_present
    MISS : usb_usual_clear_present

The ub driver is an old attempt for a slow USB Block Device driver,
it hasn't been maintained in quite a while and we don't have an easy
way to enable it.

libusual is a strange mechanism for the kernel to auto-select between
usb-storage and ub.  It works by stealing the aliases from them both
and modprobe'ing (from the kernel) the other module when a device is
connected.

The trouble is that because it's built into the kernel, on machines with
particularly fast init, it could be trying to call it before the initramfs
has even been switched to - or before the initramfs has called "depmod".
In either case, since there's no modules.alias/modules.dep files, the
modprobe call with fail and the USB storage device will not be enabled.

Plugging a second USB storage device in will magically cure the problem.

Signed-off-by: Scott James Remnant <scott@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>