ARM.imx51 Freescale:ENGR00111644 i.MX51 usb gadget mass storage data in failure
When usb gadget MSC data in(i.MX51=>PC) in INCR8 mode, USB
Controller IC need read the completed dtd after its ACTIVE
bit to be 0. if SW free the completed dtd memory before the
read, may result in data copy failure.
This is a software workaround to avoid the issue case:
when free dtds of a usb req, postpone free the last dtd. so
the dtd before an ACTIVE dtd will always be freed in next IOC
irq.

Signed-off-by: Li Jun <r65092@freescale.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>