ALSA: hdsp - poll for iobox
BugLink: https://bugs.launchpad.net/bugs/363003

commit e588ed8304f76cbb396ee85e657a58990298a675 upstream

sleeping for 2 seconds before checking for the iobox is not enough
on some systems.
this patch increases the timeout, but polls the card during that
time. it thus speeds up the module loading when the card has already
been initialized, while being more robust on systems, which require
a higher timeout than the predefined 2 seconds.

Signed-off-by: Tim Blechmann <tim@klingt.org>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Tim Gardner <tim.gardner@canonica.com>
Acked-by: Andy Whitcroft <apw@canonica.com>