ARM.imx51 Freescale:ENGR00113331 set host mode after usbcmd reset for ehci_run.
For those designs that contain both host & device capability,
the controller will default to an idle state and will need to
be initialized to the desired operating mode after reset.
For combination host/device controllers of FSL, SW need program
USBMODE register to make controller to be in host mode.

Signed-off-by: Li Jun <r65092@freescale.com>
Signed-off-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>